fix: syntax conflicts
Due to another report of this:
https://forum.sublimetext.com/t/view-syntax-open-all-with-current-extension-as-scss-please/18129

Just remove the files. They should not be there. If people do not have
the syntax definitions for these files installed, then they probably
don't have any of these files in their project anyway (or wouldn't care
about it), in which case they'll just see the generic file icon.

Closes #16